Disables TLB accesses from other memory load/store helpers
Specifically, we disable the TLB in cpu_ld and cpu_lds, which are
the memory load analogs of cpu_st.

Before this change, the load from memory in cmpxchg8b and friends
*would* still go through the TLB, so running cmpxchg8b multiple
times over the same memory region would only report a single read,
on the first and only TLB miss.