qemu/target/avr: add support for DES instruction.
Note: the DES instruction alone is not supported.

Rather, the actual DES encryption/decryption code sequence is detected
and a helper function called to execute the operation based on the H
flag.

Rationale: the Microchip instruction set manual mandates that the sixteen
rounds must be executed in increasing order. So, we expect the sequence
DES 0x0 up to including DES 0xf, without gaps.

Signed-off-by: Glenn Baker <glenn.baker@gmx.com>