Fix Python 3.8 to 3.11
Previous logic required old setuptools on py3.8 up to py3.11.
So if new setuptools was installed, importing unicorn always failed with
ImportError on these Python versions.

It now tries to use old setuptools on py3.8, but falls back gracefully
to not using anything.
So it works on py3.9+ fully, and on py3.8 it also works but with more
limited path discovery (I think it is enough, since no one complained
yet).

Fixes: 0c34496a395eac567dc24eb6574550938a52b126 ("Modify canonicals import")
Ref: https://github.com/unicorn-engine/unicorn/pull/2016/commits/0c34496a395eac567dc24eb6574550938a52b126
Ref: https://github.com/Gallopsled/pwntools/pull/2556